Check for Order Items:
Similar to the btnPayActionPerformed method, you first check if there are any items in the orderTable. If not, display a warning message.
Prepare Receipt Details:

Fetch the current date and time.
Retrieve the employee's name from the logged-in user.
Calculate the subtotal and VAT. The total is computed as the sum of the subtotal and VAT.
Use a method (like getSaleItemsFromOrderTable()) to get the list of items currently in the order table.
Display Receipt:

Create an instance of the Receipt class.
Call the displayReceipt method to populate and show the receipt. Use "N/A" or appropriate placeholder values for the sale ID and payment method since you're not completing a sale.
Set JFrame Visible:

Finally, make the receipt frame visible to the user.